vo_sharedbuffer: readd this VO
This is at least useful to protoype GUIs. I had to rip out anything related to
OSD since all the old internal APIs were ripped out of mpv when all the VOs
were made capable of rendering OSD and subs in native resolution (instead of
soldering it into the video frames).

`vf=sub` has to be added to see any kind of subtitles as it is not added by
default like mplayer and mplayer2 add `vf=ass`.

Changed a the memcpy_pic to a simple memcpy since AFAIK the stride is always
the same for source and destination and memcpy_pic was deleted from the mpv's
codebase long ago.