osc: Add internal tracklist management
The track-buttons will now work on internal tracklists to provide the necessary flexibillity. Input is always proccessed based on the currently actiuve track ID in mpv to make sure everything always stays in sync.
The iAmAProgrammer-mode will soon be modified to entirely disable the internal playlists (and functions that build on them)