input: fix mouse pointer possibly not hiding after timeout
If MOUSE_MOVE is unmapped (or mapped to something that doesn't parse),
and produces no command, the mouse position wouldn't be updated (as
the mouse position is bound to input commands). In particular, this
wouldn't update flags like whether the cursor should be hidden if the
OSC was active, and the mouse was moved into the OSC area, and then
moved outside of it (it kept using the OSC's flags).

Fix this by generating a dummy command if needed.