Replace almost all mp_msg() calls
Replace them with equivalent calls which require a context. The final
goal is to make mpv library-safe in the sense that there is no global
state.

In most cases, we have to add mp_log as fields or function arguments
in various places. In some cases, it's a bit harder and major
refactoring is done. Sometimes, just remove a "hard" mp_msg call.

xlib, alsa, ffmpeg, lcms2 all are not library-safe. Interestingly, this
all affects logging callbacks (the same problem as this commit tries to
avoid to users of a hypothetical mpv client library). Just remove the
xlib and alsa parts. For ffmpeg and lcms2 we add global locks to avoid
possible memory errors in corner cases involving multiple mpv instances
(or other library users in the same process).

Exactly 5 mp_msg() calls are still left.